Add permission hook for provider credentials
