+++ src/components/MAPI/clusters/ClusterDetail/ClusterDetailWidgetProviderAzure.tsx
@@ -15,7 +15,7 @@ import ErrorReporter from 'utils/errors/ErrorReporter';
 import { FlashMessage, messageTTL, messageType } from 'utils/flashMessage';
 import { useHttpClient } from 'utils/hooks/useHttpClient';
 
-import { usePermissionsForOrgCredentials } from '../permissions/usePermissionsForOrgCredentials';
+import { usePermissionsForProviderCredentials } from '../permissions/usePermissionsForProviderCredentials';
 import { getCredentialsAccountID, getCredentialsAzureTenantID } from './utils';
 
 const ValueWrapper = styled.div`
@@ -40,7 +40,7 @@ const ClusterDetailWidgetProviderAzure: React.FC<
 
   const provider = window.config.info.general.provider;
 
-  const { canList } = usePermissionsForOrgCredentials(
+  const { canList } = usePermissionsForProviderCredentials(
     provider,
     selectedOrg?.namespace ?? ''
   );
